hopefully solved the conversion problem :)
it's a lot of metaprogramming now but this reduces the risk getting unnecessary compiler errors:
if (convert(a, b))
  return true;
else
  return convertFallback(a, b);

if convert(a, b) is defined and always used you could still get an error if convertFallback(a, b) uses something (like stream << object) that isn't supported for a or b. but with some template tricks it's possible to not even consider using convertFallback(a, b) at compiletime and avoid the error as long as convert(a, b) is implemented.

util/Conver.h is now opened for curious coders ;)