added ClassTreeMaskObjectIterator this iterator iterates through all objects of all classes, included by a given ClassTreeMask.
Note: The ClassTreeMaskObjectIterator is fast if you only include classes. If you exclude a subclass of an included class, the performance is not that good.
If we'll use this more often, we have to think about a new implementation, using a second ObjectList with all objects of exactly one class.

Additionally, there is a small change in MultiType.h