Fixed the problem mentioned in r2362 which caused the server to crash if the client disconnected (the problem was not caused by the network but rather by the destruction of a WorldEntity).
However there is still a problem causing clients to just shut down sometimes without sending a disconnect message to the server. They are removed after some seconds because of the timeout. Try starting the server with the debugger if you can't reproduce the issue. I'm not 100% sure but I thought disconnecting worked some days or weeks ago.