Added FindLua.cmake: - Finds both version 5.0 and 5.1, but prefers 5.1 - Regards version arguments (5.1.3 EXACT will only match 5.1.3, whereas 5 EXACT will match all Lua 5 versions) - Checks the header for the right version always - Also finds debug libraries
As you might know, there are already two scripts in the CMake module path: FindLua50.cmake and FindLua51.cmake. Unfortunately it would require writing a lot of hacky extra code to meet our needs with Lua versions. So I decided to write a whole new find script that should be able to all we need.