Magician Beni on what he does best: Fixed the problem with the large scale models being too dark with only one line of change. I know everybody will love me now, so I would like you to keep relaxed and use your enthusiasm to code even greater stuff.
How great is it to finally be free using normal colored textures even when the objects is going to be scaled like crazy. No more dark space stations, no more dark planets and everybody is dancing around and holding hands. Spring has arrived and with it a brightly illuminated universe. Orxonox is so great and really the hottest (and now also the bright enough) space shooter ever to exist.

If only all problems could be solved by setting a bool to true:
world->peaceInTheMiddleEast(true);
this->usa_->noMoreAggressiveForeignPolicies(true);
eth->fiftyPercentGirls(true);