Doxygen fixes. Please advise me on better places or methods to fix it.
At the moment the Orxonox Doxygen Documentation is called "Bullet Documentation" due to Doxygen code inside the included Bullet files. I commented out this main-page of the documentation as it will most probably solve the problem.
Further I changed the version to 0.0.2 Bellatrix as this is the current version we're working on as far as I know.

Oh yeah, I added a doxygen mainpage comment at the beginning of Main.cc maybe there's a better place for this.