changed rotation convertion.
+pitch is a counterclockwise rotation around the x axis (pointing to the right)
+yaw is a counterclockwise rotation around the y axis (pointing upwards)
+roll is a counterclockwise rotation around the z axis (pointing backwards)

This corresponds to the implementation of Ogre and Bullet, so we don't have to apply magic "-" signs anymore.

Don't get misleaded by wikipedia, they use another convention using a weird mix of clockwise and counterclockwise rotation. Our system is rather like this, except for the names of the axes:
http://developer.valvesoftware.com/w/images/7/7e/Roll_pitch_yaw.gif

So a rotation around an axis is counterclockwise if the clock faces in the direction of the axis. I hope you get it. ;)

(This will break the AI even more, but it will be fixed in a later commit.)