Added support for named events (so one object can fire multiple different events). Added a new class, EventFilter, to filter differently named events and map them to different states.
I've wrapped the event names with a pretty useless macro, but I want to expand this in the future and register possible event names in the Identifier of the corresponding class. This allows us to get the possible event names in a static manner and thus build an editor upon it.