Adding new level notifications.oxw, to "showcase", or at this stage rather "test" Notifications. Restoring tutorial.oxw to its old state, before it was hijacked by me for testing.
Extending Script class. Now also normal orxonox code can be executed with it, the execution of code can be triggered with Triggers (obviously) and cod can also executed on load.
I needed this to load the NotificationLayer in levels where it is needed.
Also inserted a Script that loads the NotificationQueue to display Notifications in all levels it was needed.